requirements.txt
numpy
streamlit
plotly.graph_objects